pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Dianashafee <diana.shafee1997@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
docker = "^4.4.1"

[tool.poetry.dev-dependencies]
Django = "^3.1.5"
black = "^20.8b1"
djangorestframework = "^3.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
